pyproject.toml
[project]
name = "llm-mcp-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
        "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]
